rebase the code on refactor branch
